Applying patch from clint
