+++ processing/src/main/java/org/apache/druid/collections/spatial/ImmutableRTree.java
@@ -65,7 +65,7 @@ public final class ImmutableRTree implements Comparable<ImmutableRTree>
     Preconditions.checkArgument(data.get(initPosition) == VERSION, "Mismatching versions");
     this.numDims = data.getInt(1 + initPosition) & 0x7FFF;
     this.data = data;
-    this.root = new ImmutableNode(numDims, initPosition, 1 + Integer.BYTES, data, bitmapFactory);
+    this.root = new ImmutableFloatNode(numDims, initPosition, 1 + Integer.BYTES, data, bitmapFactory);
   }
 
   public static ImmutableRTree newImmutableFromMutable(RTree rTree)
