+++ processing/src/main/java/org/apache/druid/collections/spatial/search/RectangularBound.java
@@ -25,7 +25,7 @@ import com.google.common.base.Preconditions;
 import com.google.common.base.Predicate;
 import com.google.common.collect.Iterables;
 import org.apache.druid.collections.spatial.ImmutableNode;
-import org.apache.druid.collections.spatial.ImmutablePoint;
+import org.apache.druid.collections.spatial.ImmutableFloatPoint;
 
 import java.nio.ByteBuffer;
 import java.util.Arrays;
@@ -33,7 +33,7 @@ import java.util.Objects;
 
 /**
  */
-public class RectangularBound implements Bound
+public class RectangularBound implements Bound<float[], ImmutableFloatPoint>
 {
   private static final byte CACHE_TYPE_ID = 0x0;
 
@@ -92,7 +92,7 @@ public class RectangularBound implements Bound
   }
 
   @Override
-  public boolean overlaps(ImmutableNode node)
+  public boolean overlaps(ImmutableNode<float[]> node)
   {
     final float[] nodeMinCoords = node.getMinCoordinates();
     final float[] nodeMaxCoords = node.getMaxCoordinates();
@@ -119,14 +119,14 @@ public class RectangularBound implements Bound
   }
 
   @Override
-  public Iterable<ImmutablePoint> filter(Iterable<ImmutablePoint> points)
+  public Iterable<ImmutableFloatPoint> filter(Iterable<ImmutableFloatPoint> points)
   {
     return Iterables.filter(
         points,
-        new Predicate<ImmutablePoint>()
+        new Predicate<ImmutableFloatPoint>()
         {
           @Override
-          public boolean apply(ImmutablePoint immutablePoint)
+          public boolean apply(ImmutableFloatPoint immutablePoint)
           {
             return contains(immutablePoint.getCoords());
           }
